requirements.txt
telethon==1.34.0
flask==3.0.0
gunicorn==21.2.0
cryptg==0.4.0
google-generativeai==0.3.2
python-dotenv==1.0.0
deep-translator==1.11.4
google-genai==0.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
PySocks==1.7.1
fake-useragent==1.4.0
groq==0.4.2
beautifulsoup4==4.12.3
pillow==10.2.0
aiosqlite==0.20.0
aiohttp==3.9.3
certifi==2024.2.2
praw==7.7.1
psutil==5.9.8
